requirements.txt
Django==2.2
djangorestframework==3.9.2
django-cors-headers==3.0.2
psycopg2==2.8.2
pytz==2019.1
sqlparse==0.3.0
